+++ apps/builder/src/page/App/components/PanelSetters/ColumnSetter/utils/generateNewOptions.ts
@@ -1,9 +1,13 @@
 import { v4 } from "uuid"
 import { ColumnItemShape } from "@/widgetLibrary/TableWidget/interface"
 
-export const generateOptionItemId = () => `option-${v4()}`
+export const generateOptionItemId = () => `column-${v4()}`
 
 export const generateNewOptionItem = (number: number): ColumnItemShape => ({
   header: `Custom Column ${number}`,
-  accessorKey: generateOptionItemId(),
+  accessorKey: `Custom Column ${number}`,
+  enableSorting: true,
+  type: "text",
+  visible: true,
+  format: "YYYY-MM-DD",
 })
