+++ apps/builder/src/widgetLibrary/TableWidget/utils.ts
@@ -1,0 +1,18 @@
+import { ColumnItemShape } from "@/widgetLibrary/TableWidget/interface"
+
+export const tansTableDataToColumns = (data: Record<any, any>[]):ColumnItemShape[]  => {
+  const columns: ColumnItemShape[] = []
+  if (data && data.length > 0) {
+    Object.keys(data[0]).forEach((key) => {
+      columns.push({
+        header: key,
+        accessorKey: key,
+        enableSorting: true,
+        type: "text",
+        visible: true,
+        format: "YYYY-MM-DD",
+      })
+    })
+  }
+  return columns
+}
