+++ apps/builder/src/widgetLibrary/TableWidget/widgetConfig.tsx
@@ -1,11 +1,7 @@
 import { TextWidgetIcon } from "@illa-design/icon"
 import { WidgetConfig } from "@/widgetLibrary/interface"
 import i18n from "@/i18n/config"
-import store from "@/store"
-import { getActionList } from "@/redux/currentApp/action/actionSelector"
-import { ColumnDef } from "@tanstack/react-table"
-import { ColumnItemShape } from "@/widgetLibrary/TableWidget/interface"
-import { isNumber } from "@illa-design/system"
+import { tansTableDataToColumns } from "@/widgetLibrary/TableWidget/utils"
 
 const originData = [
   {
@@ -93,23 +89,9 @@ export const TABLE_WIDGET_CONFIG: WidgetConfig = {
 }
 
 export function initTableWidgetDefaultProps() {
-  const columns: ColumnItemShape[] = []
-  if (originData && originData.length > 0) {
-    Object.keys(originData[0]).forEach((key) => {
-      columns.push({
-        header: key,
-        accessorKey: key,
-        enableSorting: true,
-        type: "text",
-        visible: true,
-        format: "YYYY-MM-DD",
-      })
-    })
-  }
-
   return {
     data: `{{${JSON.stringify(originData)}}}`,
-    columns,
+    columns: tansTableDataToColumns(originData),
     emptyState: "No rows found",
   }
 }
